frostfact/requirements.txt
asgiref==3.8.1
boto3==1.34.139
botocore==1.34.139
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
defusedxml==0.8.0rc2
distlib==0.3.8
Django==5.0.6
django-cors-headers==4.4.0
django-filter==24.2
django-storages==1.14.3
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
filelock==3.15.4
idna==3.7
jmespath==1.0.1
Markdown==3.6
oauthlib==3.2.2
pillow==10.4.0
pipenv==2024.0.1
platformdirs==4.2.2
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
s3transfer==0.10.2
setuptools==70.2.0
six==1.16.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.26.3
whitenoise==6.7.0
